+++ wp-content/themes/oceanwp/assets/js/third/woo/devs/wooQuantityButtons.js
@@ -62,7 +62,8 @@ function oceanwpWooQuantityButtons( $quantitySelector ) {
 			// If floating bar is enabled
 			if ( $j( 'body' ).hasClass( 'single-product' )
 				&& 'on' == oceanwpLocalize.floating_bar
-				&& ! $cart.hasClass( 'grouped_form' ) ) {
+				&& ! $cart.hasClass( 'grouped_form' )
+				&& ! $cart.hasClass( 'cart_group' ) ) {
 				$quantityBox = $j( '.plus, .minus' ).closest( '.quantity' ).find( $quantitySelector );
 			} else {
 				$quantityBox = $j( this ).closest( '.quantity' ).find( $quantitySelector );